stretch child slots to fill parent, all levels
