+++ preprocessing/partial_batch.js
@@ -13,114 +13,9 @@ const isoFileName = "preprocessing/iso_dict.json";
 const indexFileName = "data/index2.json"; // stores metadata for all documents
 const dbFileName = "data/fake_database.json";
 
-//console.log("Converting all files to .JSON and re-building the database. The index and metadata will also be updated during this process. Status messages will appear below:")
+//console.log("Converting all files to .JSON and re-building the database. The index and metadata will also be updated during this process. Status messages will appear below:");
 
-const eafFileNames = ['Intro.eaf', 'flor_flanca.eaf', 'test.eaf'];
-
-/*const annotationChildren = {
-  'root':{
-    'T1':['c1a','c1b','c1c'],
-    'T2':['c2'],
-  },
-  'c2':{
-    'T4':['c4'],
-  }
-};
-
-const tiersToConstraints = {
-  'T0': '', 
-  'T1': 'Symbolic_Subdivision', 
-  'T2': 'Symbolic_Association',
-  'T4': 'Symbolic_Association'
-};
-
-const annotationsFromIDs = {
-  'root': {"ALIGNABLE_ANNOTATION": [
-      {
-        "$": {
-          "ANNOTATION_ID": "root",
-          "TIME_SLOT_REF1": "ts1",
-          "TIME_SLOT_REF2": "ts8"
-        },
-        "ANNOTATION_VALUE": [
-          "parsnip"
-        ]
-      }
-    ]
-  },
-  'c1a': {"REF_ANNOTATION": [
-      {
-        "$": {
-          "ANNOTATION_ID": "c1a",
-          "ANNOTATION_REF": "root"
-        },
-        "ANNOTATION_VALUE": [
-          "aardvark"
-        ]
-      }
-    ]
-  },
-  'c1b': {"REF_ANNOTATION": [
-      {
-        "$": {
-          "ANNOTATION_ID": "c1b",
-          "ANNOTATION_REF": "root",
-          "PREVIOUS_ANNOTATION": "c1a"
-        },
-        "ANNOTATION_VALUE": [
-          "boy"
-        ]
-      }
-    ]
-  },
-  'c1c': {"REF_ANNOTATION": [
-      {
-        "$": {
-          "ANNOTATION_ID": "c1c",
-          "ANNOTATION_REF": "root",
-          "PREVIOUS_ANNOTATION": "c1b"
-        },
-        "ANNOTATION_VALUE": [
-          "cat"
-        ]
-      }
-    ]
-  },
-  'c2': {"REF_ANNOTATION": [
-      {
-        "$": {
-          "ANNOTATION_ID": "c2",
-          "ANNOTATION_REF": "root",
-        },
-        "ANNOTATION_VALUE": [
-          "rogue"
-        ]
-      }
-    ]
-  },
-  'c4': {"REF_ANNOTATION": [
-      {
-        "$": {
-          "ANNOTATION_ID": "c4",
-          "ANNOTATION_REF": "c2",
-        },
-        "ANNOTATION_VALUE": [
-          "orphan"
-        ]
-      }
-    ]
-  },
-}
-
-const timeslots = {'ts1':0,'ts8':8000};
-
-const startSlots = {};
-const endSlots = {};
-
-elan.assignSlots('root', {contents:0}, tiersToConstraints, 
-    annotationChildren, annotationsFromIDs, timeslots, startSlots, endSlots);
-console.log(startSlots);
-console.log(endSlots);*/
+const eafFileNames = ['flor_flanca.eaf','Intro.eaf','test.eaf'];
 
 // use this to wait for things to terminate before executing the callback
 const status = {numJobs: eafFileNames.length};
